+++ CHANGELOG.md
@@ -6,6 +6,8 @@ All notable changes to [diagram-js](https://github.com/bpmn-io/diagram-js) are d
 
 _**Note:** Yet to be released changes appear here._
 
+* `FEAT`: support `imageHtml` option for popup menu header entries ([#819](https://github.com/bpmn-io/diagram-js/pull/819))
+
 ## 12.6.0
 
 * `FEAT`: support custom outline providers ([#817](https://github.com/bpmn-io/diagram-js/pull/817))